+++ Sources/KeyboardKit/_Essentials/Keyboard+State.swift
@@ -49,7 +49,10 @@ public extension Keyboard.State {
     // Setup the state instance for the provided controller.
     func setup(for controller: KeyboardInputViewController) {
         let isPhone = UIDevice.current.userInterfaceIdiom == .phone
+        let keyboardType = controller.textDocumentProxy.keyboardType
+        let contextType = keyboardType?.keyboardType ?? .alphabetic(.auto)
         keyboardContext.sync(with: controller)
+        keyboardContext.keyboardType = contextType
         calloutContext.inputContext.isEnabled = isPhone
     }
 }